task: Changed type to float

+++ 0x00-MySQL_Advanced/100-average_weighted_score.sql
@@ -4,8 +4,8 @@ DELIMITER $$
 CREATE PROCEDURE ComputeAverageWeightedScoreForUser(IN user_id INT)
 compute_average:
 BEGIN
-  DECLARE total_weight INT DEFAULT 0;
-  DECLARE project_average INT DEFAULT 0;
+  DECLARE total_weight FLOAT DEFAULT 0;
+  DECLARE project_average FLOAT DEFAULT 0;
   SELECT SUM(p.weight) INTO total_weight FROM projects p INNER JOIN corrections c ON p.id = c.project_id WHERE c.user_id = user_id;
   IF total_weight > 0 THEN
     SELECT SUM(c.score * p.weight) INTO project_average FROM corrections c INNER JOIN projects p ON c.project_id = p.id WHERE c.user_id = user_id;
